Use fba endpoint for AsaGo
