+++ mobile/asa-go/src/components/LoginButton.tsx
@@ -1,7 +1,22 @@
+import axios from "@/api/axios";
 import { authenticate } from "@/slices/authenticationSlice";
-import { AppDispatch } from "@/store";
+import { AppDispatch, AppThunk, selectToken } from "@/store";
 import { Button, Typography, useTheme } from "@mui/material";
+
+import { isNil } from "lodash";
 import { useDispatch } from "react-redux";
+
+const setAxiosRequestInterceptors = (): AppThunk => (_, getState) => {
+  // Use axios interceptors to intercept any requests and add authorization headers.
+  axios.interceptors.request.use((config) => {
+    const token = selectToken(getState());
+    if (!isNil(token)) {
+      config.headers.Authorization = `Bearer ${token}`;
+    }
+
+    return config;
+  });
+};
 
 interface LoginButonProps {
   label: string;
@@ -13,6 +28,7 @@ const LoginButton = ({ label }: LoginButonProps) => {
 
   const handleLogin = () => {
     dispatch(authenticate());
+    dispatch(setAxiosRequestInterceptors());
   };
   return (
     <Button
